+++ src/main/kotlin/cloud/glitchdev/rfu/utils/TextUtils.kt
@@ -1,6 +1,7 @@
 package cloud.glitchdev.rfu.utils
 
 import cloud.glitchdev.rfu.constants.text.TextColor
+import cloud.glitchdev.rfu.constants.text.TextEffects
 import cloud.glitchdev.rfu.constants.text.TextStyle
 import net.minecraft.network.chat.MutableComponent
 import net.minecraft.network.chat.Component
@@ -12,6 +13,10 @@ object TextUtils {
 
     fun rfuLiteral(string: String, textColor: TextColor = TextColor.WHITE) : MutableComponent {
         return Component.literal("§b§l[§f§lRFU§b§l] $textColor$string")
+    }
+
+    fun rfuLiteral(string: String, textColor: TextColor = TextColor.WHITE, textEffect : TextEffects) : MutableComponent {
+        return Component.literal("§b§l[§f§lRFU§b§l] $textColor$textEffect$string")
     }
 
     fun rfupfLiteral(string: String, textStyle: TextStyle) : MutableComponent {